UnitTsting/UnitTsting/Person.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace UnitTsting
{
    public class Person
    {
        public string Firstname { get; set; }
        public string Lastname { get; set; }
        public int Age { get; set; }

        public Person(string firstname, string lastname, int age)
        {
            Firstname = firstname;
            Lastname = lastname;
            Age = age;
        }

        public bool CanVote()
        {
            return Age <= 18;
        }

        public int StringLength(string name)
        {
            return name.Length ;
        }
    }
}
